add new validation directory for validation and error utilities
